requirements.txt
yt-dlp>=2025.0
youtube-transcript-api>=1.0
openai-whisper>=2025.0
torch>=2.2
requests>=2.31
